refactor(agent): move KiroBackend into its own kiro_backend.py
Review feedback on the pull request asked for the kiro code to live in
its own file, mirroring the claude backend, so that backend.py stays a
pure extension point. KiroBackend moves unchanged into
cve_agent/kiro_backend.py, named consistently with claude_backend.py;
backend.py now holds only the AIBackend interface, SessionResult, the
registry, and plugin discovery.

Both built-in backends now register the same way - lazily, on first
use, via _ensure_builtin_backends() - so kiro gets the same
import-order guarantee the claude module already had, covered by a new
fresh-interpreter regression test. A plugin registered under the same
name before first use still wins, preserving extra/ override semantics.

No behavioural change: the class body is diff-identical to the old
backend.py block. Test importers and the architecture docs are updated
for the new module path.

Re-verified after the move: ruff and mypy clean, 719 tests pass with
75% coverage, live claude smoke tests pass (2/2), and integration case
11 (CVE-2024-39894 / openssh, AGENT_BACKEND=claude) passes end to end
on a real scarthgap build.
