+++ cve_agent/backend.py
@@ -3,15 +3,9 @@
 """Pluggable AI backend interface for CVE agent sessions."""
 import logging
 import os
-import subprocess
-import time
 from dataclasses import dataclass
 from pathlib import Path
 from typing import Optional
-
-from shared import build_git_env
-
-from .git import has_in_progress_operation
 
 
 @dataclass
@@ -44,64 +38,7 @@ class AIBackend:
         """Perform any one-time setup."""
 
 
-class KiroBackend(AIBackend):
-    """Default backend using kiro-cli."""
-    name = "kiro"
-
-    def is_available(self) -> bool:
-        import shutil
-        return shutil.which("kiro-cli") is not None
-
-    def run_session(self, prompt: str, workspace_path: Path,
-                   allowed_files: set, model: str,
-                   timeout: int, interactive: bool) -> SessionResult:
-        agent_name = ('yocto-cve-backport-interactive' if interactive
-                      else 'yocto-cve-backport')
-        env = build_git_env()
-        cmd = ['kiro-cli', 'chat', '--agent', agent_name, '--model', model]
-        if not interactive:
-            cmd.append('--no-interactive')
-            cmd.append('--trust-tools=fs_read,fs_write,execute_bash')
-        cmd.append(prompt)
-
-        start = time.monotonic()
-        timed_out = False
-        try:
-            subprocess.run(cmd, cwd=workspace_path, env=env,
-                         check=False, timeout=timeout)
-        except subprocess.TimeoutExpired:
-            timed_out = True
-        except FileNotFoundError:
-            logging.error("kiro-cli not found. Install it or add to PATH.")
-        except KeyboardInterrupt:
-            pass
-
-        duration = time.monotonic() - start
-        if timed_out:
-            return SessionResult(resolved=False, duration=duration)
-
-        resolved = self._check_resolution(workspace_path)
-        return SessionResult(resolved=resolved, duration=duration)
-
-    def _check_resolution(self, workspace_path: Path) -> bool:
-        if has_in_progress_operation(workspace_path):
-            return False
-        result = subprocess.run(
-            ['git', 'status', '--porcelain'],
-            cwd=workspace_path, capture_output=True, text=True, check=False)
-        if result.returncode != 0:
-            return False
-        for line in result.stdout.splitlines():
-            if line and len(line) >= 2 and (line[0] == 'U' or line[1] == 'U'):
-                return False
-        return True
-
-    def setup(self, **kwargs) -> None:
-        from .setup import ensure_agents
-        ensure_agents(interactive=kwargs.get('interactive', True))
-
-
-_BACKENDS: dict[str, AIBackend] = {"kiro": KiroBackend()}
+_BACKENDS: dict[str, AIBackend] = {}
 
 
 def _ensure_builtin_backends() -> None:
@@ -111,7 +48,13 @@ def _ensure_builtin_backends() -> None:
     AIBackend/SessionResult from here — an import at the bottom of this
     module makes ``import cve_agent.claude_backend`` fail with a circular
     import whenever it is imported before ``cve_agent.backend``.
+
+    A backend already registered under the same name (an ``extra/`` plugin
+    loaded first) is left in place, so plugin override semantics hold.
     """
+    if "kiro" not in _BACKENDS:
+        from .kiro_backend import KiroBackend
+        _BACKENDS["kiro"] = KiroBackend()
     if "claude" not in _BACKENDS:
         from .claude_backend import ClaudeBackend
         _BACKENDS["claude"] = ClaudeBackend()
